more status, includes to do
TODO:
     * 1. trymerge
     * 2. finish sync
     * 3. update status ui
     * 4. add player draw
     * 5. add non-player draw (if separate)
     * 6. add full sync

+++ MPlayer.cs
@@ -1079,7 +1079,7 @@ namespace ExperienceAndClasses {
         /*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Status ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 
         public bool HasStatus(Systems.Status.IDs id) {
-            return (Status[(uint)id] == null);
+            return (Status[(short)id] == null);
         }
 
         public void UpdateStatus() {
